move remaining method from snuba/web/rpc/v1/resolvers/R_eap_items/common/common to its only use case
